Use difflib for diffing asm
This lets us align asm based on mnemonic/symbol references, and then
compare registers etc. manually after that. (The original code tried to
do roughly the same by adding whitespaces intelligently, but the use of
wdiff meant we sometimes got diff formats we didn't expect, and it was
always something of a hack.)